+++ apps/OrderProcessor/src/main/java/com/tradingplatform/orderprocessor/database/DatabaseUtils.java
@@ -6,8 +6,6 @@ import com.mongodb.client.model.Projections;
 import com.mongodb.client.model.Updates;
 import com.mongodb.client.result.UpdateResult;
 import com.tradingplatform.orderprocessor.orders.Order;
-import com.tradingplatform.orderprocessor.orders.OrderType;
-import com.tradingplatform.orderprocessor.orders.Ticker;
 import io.vertx.core.json.JsonObject;
 import java.util.ArrayList;
 import java.util.List;
@@ -15,19 +13,6 @@ import org.bson.Document;
 import org.bson.conversions.Bson;
 
 public class DatabaseUtils {
-
-  public static Order createOrder(JsonObject body) {
-    String typeStr = body.getString("type");
-    String tickerStr = body.getString("ticker");
-    double price = body.getDouble("price");
-    int quantity = body.getInteger("quantity");
-    String userId = body.getString("userId");
-
-    OrderType type = OrderType.valueOf(typeStr);
-    Ticker ticker = Ticker.valueOf(tickerStr);
-
-    return new Order(type, userId, ticker, price, quantity);
-  }
 
   public static boolean previouslyPartiallyFilled(String orderId) {
     MongoCollection<Document> orderHistoryCollection =
@@ -40,11 +25,11 @@ public class DatabaseUtils {
     return previouslyPartiallyFilled != null;
   }
 
-  public static String passValidations(
-    JsonObject body,
-    MongoCollection<Document> usersCollection
-  ) {
+  public static String passValidations(JsonObject body) {
     //TODO SPLIT INTO METHODS
+
+    MongoCollection<Document> usersCollection =
+      MongoClientConnection.getCollection("users");
 
     String typeStr = body.getString("type");
     boolean isBuy = typeStr.equals("BUY");
@@ -120,19 +105,16 @@ public class DatabaseUtils {
     return "PASSED VALIDATIONS";
   }
 
-  public static void insertOrderIntoDatabase(
-    Order order,
-    MongoCollection<Document> activeOrdersCollection
-  ) {
+  public static void insertOrderIntoDatabase(Order order) {
+    MongoCollection<Document> activeOrdersCollection =
+      MongoClientConnection.getCollection("activeOrders");
+
     Document orderDoc = order.toDoc();
     activeOrdersCollection.insertOne(orderDoc);
   }
 
-  public static void updateDb(
-    ArrayList<String> matchesFound,
-    MongoCollection<Document> activeOrdersCollection,
-    MongoCollection<Document> usersCollection,
-    MongoCollection<Document> orderHistoryCollection
+  public static void updateCollectionsWithMatches(
+    ArrayList<String> matchesFound
   ) {
     ArrayList<Document> matchesFoundAsMongoDBDocs = convertMatchesToDocs(
       matchesFound
@@ -141,20 +123,8 @@ public class DatabaseUtils {
     Document buyOrder = matchesFoundAsMongoDBDocs.get(0);
     Document sellOrder = matchesFoundAsMongoDBDocs.get(1);
 
-    updateDbAccordingToPartiallyFilledOrNot(
-      buyOrder,
-      true,
-      activeOrdersCollection,
-      usersCollection,
-      orderHistoryCollection
-    );
-    updateDbAccordingToPartiallyFilledOrNot(
-      sellOrder,
-      false,
-      activeOrdersCollection,
-      usersCollection,
-      orderHistoryCollection
-    );
+    updateDbAccordingToPartiallyFilledOrNot(buyOrder, true);
+    updateDbAccordingToPartiallyFilledOrNot(sellOrder, false);
   }
 
   private static ArrayList<Document> convertMatchesToDocs(
@@ -172,11 +142,17 @@ public class DatabaseUtils {
 
   public static void updateDbAccordingToPartiallyFilledOrNot(
     Document order,
-    boolean isBuy,
-    MongoCollection<Document> activeOrdersCollection,
-    MongoCollection<Document> usersCollection,
-    MongoCollection<Document> orderHistoryCollection
+    boolean isBuy
   ) {
+    MongoCollection<Document> activeOrdersCollection =
+      MongoClientConnection.getCollection("activeOrders");
+
+    MongoCollection<Document> usersCollection =
+      MongoClientConnection.getCollection("users");
+
+    MongoCollection<Document> orderHistoryCollection =
+      MongoClientConnection.getCollection("orderHistory");
+
     boolean filled = order.getBoolean("filled");
     String orderId = order.getString("orderID");
     String userId = order.getString("userId");
